requirements.txt
Django==2.2.10
djangorestframework==3.9.2
pytz==2020.4
sqlparse==0.3.0
